+++ packages/cli/src/run-test.ts
@@ -4,29 +4,36 @@ import { ProjectOptions } from '@bigtest/project';
 import { Client } from '@bigtest/client';
 import { MainError } from '@effection/node';
 import * as query from './query';
-import { Formatter } from './format-helpers';
+import { Formatter, FormatterConstructor } from './format-helpers';
 
 import checks from './formatters/checks';
 import lines from './formatters/lines';
 
 interface Options {
-  formatterName?: string;
+  formatterName: string;
+  files: string[];
   showFullStack: boolean;
   showLog: boolean;
 }
 
-const BUILTIN_FORMATTERS: Record<string, Formatter> = { checks, lines };
+const BUILTIN_FORMATTERS: Record<string, Formatter | FormatterConstructor> = { checks, lines };
 
-function *resolveFormatter(name?: string): Operation<Formatter> {
-  if(!name) {
-    return checks;
+function initFormatter(input: Formatter | FormatterConstructor): Formatter {
+  if(typeof(input) === 'function') {
+    return input();
+  } else {
+    return input;
   }
+}
+
+function *resolveFormatter(name: string): Operation<Formatter> {
   let builtin = BUILTIN_FORMATTERS[name];
   if(builtin) {
-    return builtin;
+    return initFormatter(builtin);
   } else {
     try {
-      return yield import(name);
+      let imported: Formatter | FormatterConstructor = yield import(name);
+      return initFormatter(imported);
     } catch {
       throw new MainError({ exitCode: 1, message: chalk.red(`ERROR: Formatter with module name ${JSON.stringify(name)} not found.`) });
     }
@@ -52,6 +59,7 @@ export function* runTest(config: ProjectOptions, options: Options): Operation<vo
   };
 
   let subscription = yield client.subscription(query.run(), {
+    files: options.files,
     showDependenciesStackTrace: options.showFullStack,
     showInternalStackTrace: options.showFullStack,
     showStackTraceCode: options.showFullStack,