+++ packages/globals/src/index.ts
@@ -1,11 +1,15 @@
 import './globals';
 import { TestImplementation } from '@bigtest/suite';
 
+export type RunnerState = 'pending' | 'step' | 'assertion';
+
 interface BigtestOptions {
   testFrame?: HTMLIFrameElement;
   document?: Document;
   defaultInteractorTimeout?: number;
+  defaultAppTimeout?: number;
   appUrl?: string;
+  runnerState?: RunnerState;
 }
 
 const defaultManifest: TestImplementation = {
@@ -52,12 +56,28 @@ export const bigtestGlobals = {
     options().defaultInteractorTimeout = value;
   },
 
+  get defaultAppTimeout(): number {
+    return options().defaultAppTimeout || 20000;
+  },
+
+  set defaultAppTimeout(value: number) {
+    options().defaultAppTimeout = value;
+  },
+
   get testFrame(): HTMLIFrameElement | undefined {
     return options().testFrame;
   },
 
   set testFrame(value: HTMLIFrameElement | undefined) {
     options().testFrame = value;
+  },
+
+  get runnerState(): RunnerState | undefined {
+    return options().runnerState;
+  },
+
+  set runnerState(value: RunnerState | undefined) {
+    options().runnerState = value;
   },
 
   get appUrl(): string | undefined {