+++ packages/agent/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @bigtest/agent
 
+## 0.13.0
+
+### Minor Changes
+
+- c7bed38b: Allow step actions and assertion checks to return values synchronously
+- 7a727c86: Prevent interactor actions from being run in assertions
+
+### Patch Changes
+
+- eae589f0: Hide splash page when app frame is active
+- Updated dependencies [d97038a8]
+- Updated dependencies [7a727c86]
+  - @bigtest/globals@0.7.0
+
+## 0.12.0
+
+### Minor Changes
+
+- f1ed2b61: remove all sessionStorage, localStorage, and indexedDB data before
+  each and every test run in order to enforce test isolation.
+
 ## 0.11.0
 
 ### Minor Changes